Initial checkin.
This python script will take a series of files specified by a pattern
and turn them into a tile set file and a Tiled (.tmx) file that
references it.  The Tiled file will have a layer for each of the
original imput files, named after the input file.

The script will attempt to remove duplicate images by comparing the
eight principal rotations supported by Tiled (flipX, flipY, flipD, in
all combinations).  The script looks for an exact match on a rotation,
so even a single value difference in a single pixel can cause two
images to be different.  It also removes empty tiles completely.

Because of the way tile maps are constructed, empty tiles make up the
majority of data in all but a few layers.  The script creates an
“empty” tile as part of the set and tries to match any new tiles found
to this immediately to cull the new tile if it is empty.

New tiles are immediately checked against existing tiles for a match
when they are added.  The bookkeeping becomes more complex if you find
all the possible tiles and then try to find the duplicates.

The program has options to control the input source pattern as well as
the names (and path) to the output files.  However, these are not used
on the command line, yet, so this will be added.

Also, the program overwrites an existing Tiled file if it finds it.
This is impractical for situations where you create your map, create
your tiles, then try to modify it and recreate your Tiled file.  If you
have modified the Tiled file, you will lose your changes.  And there
can be A LOT of changes.  Some sort of intelligent compromise will have
to be made.

For now, the proposed solution (not implemented yet):
1.  A flag will control an attempt to merge or not.  No Merge =
Overwrite immediately.
Assuming the “merge” flag is set:
2. If the dimensions (tile width or height, or map width/height) are
different abort with a message.
3. If the tiles set is the same size, preserve all properties in the
original set.  Otherwise abort with a message.
4. Preserve any object layers in the original.
5. For any layers in the new and the original, preserve any properties
from the original.
6. For any layers in the original but not in the new, preserve them
entirely.