Added option parser to the script.
I was originally using parsearg, but it got complicated.  After a
little search, I found docopt.  I found it a bit easier to use, though
it must be included with he project (less than optimal, but not a big
deal).

The MapTiler itself works pretty well so far in my tests.