Added several fixes and optimizations.
- Added guards for non-square tiles.
- Added handling of images that do not have an RGBA format.
- Added optimization where the current tile is checked against the last
tile first.  If it is , just use it.  If not, start comparison against
existing tiles.
- Added option for verbose output.
- Added option to select a subimage of the processed layers.
- Added percent complete in verbose output.
- I did some experiments with estimating the time remaining.  However,
this is NOT a linear function of how far you have come, but probably a
square law random function.  This is because the search time for new
tiles gets longer each time you add a new tile.  So instead of trying
to estimate it, I am just leaving it out for now.