Much better.
Added some code to segment the map into rooms based on a combination of
a  “Doors” layer and a “Portals” layer.  The former are drawn doors.
The latter can be used for sections where the area needs to be
segmented for other regions but an explicit door may be confusing.