Modifying the model slightly.
Instead of having a single objects layer, I believe the workflow will
go smoother if I have a map layer for each type of object I want to
represent.  So the “Healing” object will have its own layer, “Weapons”
have their own layer, etc.  This is a simple game (with a single
weapon).  The problem is that maps get reworked a lot and having all
the tags in the modified .tmx file getting merged each time is a pain.

So the modified flow is:
1. Create the map in a vector/art editor.
2. Run the script to generate layers.
3. Verify the map looks good in Tiled.
4. Use secondary processing (different script) to generate graphs,
objects, validation, etc.

The point is to lose as little work as possible when you go from
modifying the map to getting your data into your game.

Some advantages:
1. Adding a new object to the map is easy, because you add it in the
graphics editor.
2. When creating new maps, copying a layer and then re-using it is very
easy.  The editors support this inherently.
3. You don’t put information in via Tiled so you don’t lose information
in Tiled when you modify your map.

The practical reality, at least for the work flow I am going through
now, is that changing your maps happens a lot, so you have to
streamline the flow from that point forward.