Added merge for rooms.
- Added changes to fix merging in rooms when updating a .tmx file.
- Added a True Type Font for the room numbering and centered the text.

The basic operation of the script is about where I want it to be
without embedding too much of the overall workflow into THIS script.
In general, I think a second script should be run on the .tmx file to
do any customization specific to the final usage.

This means Tiled can be a “check” tool to verify the image import
worked as expected, but that the next stage in the pipeline is used for
graph/data generation.  Or it could mean that work is done in Tiled and
there is some risk (maybe small) that the map will need an update but
not affect the Tiled entered data.