Export Data Completed
I have finished (for now) the export of the main data that was derived
from the Tiled file.

The exported file follows a strategy of two textual keys (primary,
second) followed by parameters for that key.  This allows a mapping
from the keys to enumerations and then a simple two level switch to
handle the remaining items in the row by a specific function call.
This should make for a very fast parse/load operation.

This particular format is specific to the flow that I intend to use for
the next phase, using the tilemap in a small “deathmatch” simulation
for GOAP.  So this will be the last entry for a while, save minor
updates, while that is in work.